% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{cleanNestedList}
\alias{cleanNestedList}
\title{cleanNestedList extracts sublists from station metric responses}
\usage{
cleanNestedList(l)
}
\arguments{
\item{l}{response (list) from getWxObservations() request}
}
\description{
cleanNestedList extracts sublists from station metric responses
}
