% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{formatRequest}
\alias{formatRequest}
\title{formatRequest generates the JSON-formatted request to send to ACIS}
\usage{
formatRequest(requestType, climateParameters, sdate, edate, cUid = NULL,
  duration = NULL, interval = NULL, paramFlags = NULL,
  reduceList = NULL, maxMissing = NULL, normal = NULL,
  gridElements = NULL)
}
\arguments{
\item{requestType}{type of request: getDailyWxObservations, getMonthlyWxObservations, getGrids, (findStation)}

\item{climateParameters}{A list of one or more climate parameters defined in calling source}

\item{sdate}{sdate (required) Start data defined in calling source}

\item{edate}{sdate (required) End date defined in calling source}

\item{cUid}{(optional) station UID defined in calling source, used for getWXObservation requests}

\item{duration}{(optional) station data duration specified in calling source; used for getWxObservations and getGrids requests}

\item{interval}{(optional) time interval for results specified in calling source; used for getWxObservations and getGrids requests}

\item{paramFlags}{(optional) used for getWxObservations (daily). Parameter flags: f = ACIS flag, s = source flag}

\item{reduceList}{(optional) used for getWxObservations (monthly). Defaults to min, max, sum, and mean.}

\item{maxMissing}{(optional) used for getWxObservations (monthly). Defaults to 1 which is approx 3.3 missing days per month.}

\item{normal}{(optional) 1 = 30-year climate normal or "departure" for departure from 30-year climate normal. Normal period: 1981-2010.}

\item{gridElements}{grid request values defined in calling source}
}
\description{
Requests are made via GET (findStation) or POST (getWxObservations, getGrids)
}

