% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{formatWxObservations}
\alias{formatWxObservations}
\title{formatWxObservations converts get*WxObservation response to a data frame, iterating by date and value}
\usage{
formatWxObservations(rList, duration, climateParameters, reduceCodes,
  luElements)
}
\arguments{
\item{duration}{station data duration specified in calling source}

\item{climateParameters}{A list of one or more climate parameters defined in calling source}

\item{reduceCodes}{A list of one or more reduce codes defined in calling source}

\item{luElements}{lookup values defined in calling source}

\item{responseContent}{list of response arrays containing name/value pairs: meta (default), data (date, values)}
}
\description{
formatWxObservations converts get*WxObservation response to a data frame, iterating by date and value
}

