% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{getAOAFeature}
\alias{getAOAFeature}
\title{getAOAFeature}
\usage{
getAOAFeature(unitCode, aoaExtent = "km30")
}
\arguments{
\item{unitCode}{unitCode One NPS unit code as a string}

\item{aoaExtent}{aoaExtent one of park, km3 or km30 as a string. Default is "km30"}
}
\description{
Function retrieves a GeoJSON-formatted area of analysis (AOA) polygon in the 
NAD83 geographic coordinate reference system (CRS).
}

