% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{getBBox}
\alias{getBBox}
\title{getBBox retrieves bounding box from IRMA Unit service and buffers it by specified distance}
\usage{
getBBox(unitCode, expandBBox, bboxCustom = NULL)
}
\arguments{
\item{unitCode}{unitCode One NPS or FWS unit code as a string; if matches unit service, will default to standard bounding box for unit, otherwise will use custom bounding box}

\item{expandBBox}{buffer distance in decimal degrees (assumes WGS84)}

\item{bboxCustom}{custom bounding box in format: "xmin, ymin, xmax, ymax". Example: "-114.291153779, 35.5612153111, -111.252315168, 37.0351548001"}
}
\description{
getBBox retrieves bounding box from IRMA Unit service and buffers it by specified distance
}
