% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{getDepartureCounts}
\alias{getDepartureCounts}
\title{getDepartureCounts}
\usage{
getDepartureCounts(
  rawDepartures,
  duration = "yly",
  metric = NULL,
  filePathAndName = NULL
)
}
\arguments{
\item{rawDepartures}{Daily departures for a climate parameter (use getWxObservations() with normal="departure" to generate)}

\item{duration}{Duration of summarization period. Default is yearly ("yly"). Use "mly" for monthly.}

\item{metric}{(optional) One climate metric from the IMD Environmental Setting protocol}

\item{filePathAndName}{(optional) File path and name including extension for output CSV file}
}
\value{
A data frame of stations, dates, and their departure counts above or below the 30-year climate normal (1981-2010)
}
\description{
Function calculates day counts by year or month for the station-based above and below normal metrics (CST8 and 9; CSP7 and 8) of the IMD Environmental Setting Protocol
}
