% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{getGridStatistics}
\alias{getGridStatistics}
\title{getGridStatistics}
\usage{
getGridStatistics(rasStack, metric = NULL, unitCode = NULL, filePath = NULL)
}
\arguments{
\item{rasStack}{(required) RasterStack object}

\item{metric}{(optional) One climate metric from the IMD Environmental Setting protocol}

\item{unitCode}{(optional) One unit code as a string}

\item{filePath}{(optional) full file path for CSV output}
}
\description{
Generates a CSV file containing raster statistics from a raster stack.
}
