% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{getMetricGrids}
\alias{getMetricGrids}
\title{getMetricGrids}
\usage{
getMetricGrids(
  featurePolygon,
  metric,
  unitCode,
  sdate = NULL,
  edate = NULL,
  outputNormals = FALSE,
  filePath = NULL
)
}
\arguments{
\item{featurePolygon}{SpatialPolygon object; for area of analysis, use getAOAFeature()}

\item{metric}{(required) One climate metric from the IMD Environmental Setting protocol}

\item{unitCode}{(required) unitCode One NPS unit code as a string}

\item{sdate}{sdate (required) Format as a string (yyyy-mm, yyyymm, yyyy). The beginning of the desired date range.}

\item{edate}{edate (optional) Format as a string (yyyy-mm, yyyymm, yyyy). The end of the desired date range. If missing, defaults to end of previous calendar year.}

\item{outputNormals}{outputNormals (optional) Output GeoTIFFs of monthly 30-yr normals. Defaults to FALSE.}

\item{filePath}{(optional) full file path for raster output}
}
\description{
\emph{NPS intranet only}: for CONUS areas, function retrieves cropped (clipped) 800m PRISM rasters (LT81m) from shared drive.
Source 800m rasters are used to generate metric grids (precipitation: percent of 30-yr normal, 
temperature: departure from normal).
From the PRISM documentation for the LT81m source: "Emphasis is on long-term consistency, and uses only station networks having at least
some stations with ≥ 20 years of data." 
License for source LT81m 800m PRISM data precludes outputting cropped source rasters.
}
\details{
\emph{Do not use for multi-decadal trend calculations.} 
See the \href{http://prism.oregonstate.edu/documents/PRISM_datasets.pdf}{PRISM datasets} document for more details.

Requires rgdal library and NPS intranet connection.

For the AOA, returns a raster stack containing metric layers (cropped, masked) and writes out a CSV file of metric raster statistics and PNG-formatted plots for metric, 
cropped PRISM 800m source and, optionally, cropped 30 year normal rasters.
}
