% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{getNPSPRISM}
\alias{getNPSPRISM}
\title{getNPSPRISM}
\usage{
getNPSPRISM(featurePolygon, metric, unitCode, sdate = NULL, edate = NULL,
  filePath = NULL)
}
\arguments{
\item{featurePolygon}{SpatialPolygon object; for area of analysis, use getAOAFeature()}

\item{metric}{(required) One climate metric from the IMD Environmental Setting protocol}

\item{unitCode}{(required) unitCode One NPS unit code as a string}

\item{sdate}{sdate (required) Format as a string (yyyy-mm, yyyymm, yyyy). The beginning of the desired date range.}

\item{edate}{edate (optional) Format as a string (yyyy-mm, yyyymm, yyyy). The end of the desired date range. If missing, defaults to end of previous calendar year.}

\item{filePath}{(optional) full file path for raster output}
}
\description{
NPS only: function retrieves cropped (clipped) 800m PRISM rasters from shared drive.
License for source 800m PRISM data precludes outputting cropped source rasters.
Requires rgdal library.
}
\details{
For the AOA, outputs raster stack containing metric layers (cropped, masked) and writes out metric and cropped 30 year normal GeoTIFF files.
}

