% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{getProtocolStations}
\alias{getProtocolStations}
\title{getProtocolStations}
\usage{
getProtocolStations(dbInstance, dbName, dbTable)
}
\arguments{
\item{dbInstance}{database server and instance, for example INPNISCVDBNRSST\\\\IMDGIS (note double back slash)}

\item{dbName}{database name}

\item{dbTable}{table name containing monitoring locations}
}
\value{
A list of station UIDs
}
\description{
NPS Only: Function retrieves climate station monitoring locations used to request station-based metrics of the IMD Environmental Setting Protocol
}
