% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{getRunCounts}
\alias{getRunCounts}
\title{getRunCounts}
\usage{
getRunCounts(rawCounts, runLength, metric = NULL, filePathAndName = NULL)
}
\arguments{
\item{rawCounts}{run counts for a climate parameter (use getWxObservations() with a reduce code of 'run*' to generate)}

\item{runLength}{number of run days used to filter raw run counts}

\item{metric}{(optional) One climate metric from the IMD Environmental Setting protocol}

\item{filePathAndName}{(optional) File path and name including extension for output CSV file}
}
\value{
A data frame of stations, dates, and their run counts greater than or equal to the specified runLength
}
\description{
Summarizes raw run response into a data frame with year and count of runs >= specified # of days
}

