% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R, R/utilityFunctionsCorrected.R
\name{getStationSubtype}
\alias{getStationSubtype}
\title{Utility functions for EnvironmentalSetting_Toolkit}
\usage{
getStationSubtype(testType, testSid)

getStationSubtype(testType, testSid)
}
\arguments{
\item{testType}{station identifier type code}

\item{testSid}{first three characters of the station identifier}

\item{testType}{station identifier type code}

\item{testSid}{first three characters of the station identifier}
}
\description{
getStationType function uses the station identifier type code to lookup the station type description

getStationType function uses the station identifier type code to lookup the station type description
}

