% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R, R/utilityFunctionsCorrected.R
\name{getUSHCN}
\alias{getUSHCN}
\title{getUSHCN retrieves the list of USHCN (U.S. Historical Climatology Network) station identifiers and compares that to the set of stations requested. Matches are returned as a vector with flag values (N = not HCN, Y = HCN).}
\usage{
getUSHCN(responseList)

getUSHCN(responseList)
}
\arguments{
\item{responseList}{list of response array of requested station codes (sid)}

\item{responseList}{list of response array of requested station codes (sid)}
}
\description{
getUSHCN retrieves the list of USHCN (U.S. Historical Climatology Network) station identifiers and compares that to the set of stations requested. Matches are returned as a vector with flag values (N = not HCN, Y = HCN).

getUSHCN retrieves the list of USHCN (U.S. Historical Climatology Network) station identifiers and compares that to the set of stations requested. Matches are returned as a vector with flag values (N = not HCN, Y = HCN).
}

