% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{outputAscii}
\alias{outputAscii}
\title{outputAscii formats grid(s) as ASCII (*.asc) with headers and projection (*.prj)}
\usage{
outputAscii(gridResponse, fullFilePath, lonCen, latCen, luSource)
}
\arguments{
\item{gridResponse}{grid (dataframe format) returned from ACIS request (by date)}

\item{lonCen}{longitude of lower left grid cell}

\item{luSource}{ACIS lookup source (as dataframe)}

\item{filePath}{full file path for ASCII output}

\item{lonCen}{latitude of lower left grid cell}
}
\description{
outputAscii formats grid(s) as ASCII (*.asc) with headers and projection (*.prj)
}

