% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{outputMetricFile}
\alias{outputMetricFile}
\title{outputMetricFile writes metric data frames to a CSV file}
\usage{
outputMetricFile(metricData, metricName, filePathAndRootName)
}
\arguments{
\item{metricData}{Data frame to output}

\item{metricName}{Metric name}

\item{filePathAndRootName}{Output path and folder}
}
\description{
outputMetricFile writes metric data frames to a CSV file
}
