% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{formatWxObservations}
\alias{formatWxObservations}
\title{formatWxObservations converts get*WxObservation response to a data frame, iterating by date and value}
\usage{
formatWxObservations(
  rList,
  duration,
  climateParameters,
  reduceCodes,
  luElements,
  normal,
  metric
)
}
\arguments{
\item{rList}{list of response arrays containing name/value pairs: meta (default), data (date, values)}

\item{duration}{station data duration specified in calling source}

\item{climateParameters}{A list of one or more climate parameters defined in calling source}

\item{reduceCodes}{A list of one or more reduce codes defined in calling source}

\item{luElements}{lookup values defined in calling source}

\item{normal}{(optional) 1 = 30-year climate normal or "departure" for departure from 30-year climate normal. Normal period: 1981-2010.}

\item{metric}{(optional) Metric code for IMD Environmental}
}
\description{
formatWxObservations converts get*WxObservation response to a data frame, iterating by date and value
}
