% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{getProtocolStations}
\alias{getProtocolStations}
\title{getProtocolStations}
\usage{
getProtocolStations(dbInstance, dbName, dbTable,
  withPeriodOfRecord = FALSE)
}
\arguments{
\item{dbInstance}{database server and instance, for example INPNISCVDBNRSST\\\\IMDGIS (note double back slash)}

\item{dbName}{database name}

\item{dbTable}{table name containing monitoring locations}

\item{withPeriodOfRecord}{boolean indicating if period of record by climate parameter is returned; default is FALSE}
}
\value{
A list of station UIDs or a list with station UIDs, climate parameter(s), and periods of record
}
\description{
NPS Only: Function retrieves climate station monitoring locations used to request station-based metrics of the IMD Environmental Setting Protocol
}
