% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{stripEscapes}
\alias{stripEscapes}
\title{stripEscapes strips escape characters from input string}
\usage{
stripEscapes(inputStr)
}
\arguments{
\item{inputStr}{input from which escape characters are to be stripped}
}
\description{
stripEscapes strips escape characters from input string
}
